backend/src/ai/requirements.txt
tensorflow-cpu==2.15.0
keras==2.15.0
pillow
numpy
flask
flask-cors
h5py==3.9.0